chore: save restored WIP from Task 5 stash
