+++ crates/katana-platform/src/lib.rs
@@ -1,9 +1,10 @@
-#![deny(
-    warnings,
-    clippy::all,
+#![deny(warnings, clippy::all)]
+#![allow(
+    missing_docs,
+    clippy::missing_errors_doc,
     clippy::too_many_lines,
-    clippy::cognitive_complexity,
-    clippy::wildcard_imports,
+    clippy::expect_used,
+    clippy::indexing_slicing,
     clippy::unwrap_used,
     clippy::panic,
     clippy::todo,
@@ -14,14 +15,6 @@
     clippy::indexing_slicing,
     clippy::missing_errors_doc
 )]
-#![allow(missing_docs)]
-#![allow(clippy::too_many_lines)]
-#![allow(clippy::missing_errors_doc)]
-#![allow(clippy::indexing_slicing)]
-#![allow(clippy::cognitive_complexity)]
-#![allow(clippy::unwrap_used)]
-#![allow(clippy::expect_used)]
-#![allow(clippy::panic)]
 
 pub mod cache;
 pub mod filesystem;
