+++ crates/katana-ui/src/preview_pane/tests.rs
@@ -1,0 +1,965 @@
+use super::*;
+
+#[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::field_reassign_with_default)]
+mod tests {
+    use super::*;
+    use katana_core::markdown::diagram::{DiagramBlock, DiagramKind, DiagramResult};
+
+    use katana_platform::CacheFacade;
+
+    // ── Markdown image parsing (test-only utilities) ──
+
+    /// Byte length of the badge image prefix `[![`.
+    const BADGE_PREFIX_LEN: usize = "[![".len();
+
+    /// Parsed Markdown image reference.
+    #[derive(Debug, Clone, PartialEq, Eq)]
+    struct MdImage {
+        src: String,
+        alt: String,
+        /// Number of characters consumed from the input string.
+        consumed: usize,
+    }
+
+    /// Finds the byte offset of the next image pattern (`![` or `[![`).
+    fn find_next_image(s: &str) -> Option<usize> {
+        let pos = s.find("![")?;
+        if pos > 0 && s.as_bytes()[pos - 1] == b'[' {
+            Some(pos - 1)
+        } else {
+            Some(pos)
+        }
+    }
+
+    /// Parses a Markdown image at the start of the given string.
+    fn parse_md_image(s: &str) -> Option<MdImage> {
+        if let Some(rest) = s.strip_prefix("[![") {
+            let alt_end = rest.find(']')?;
+            let alt = &rest[..alt_end];
+            let after_alt = &rest[alt_end + 1..];
+            let inner_src = after_alt.strip_prefix('(')?;
+            let src_end = inner_src.find(')')?;
+            let src = &inner_src[..src_end];
+            let after_inner = &inner_src[src_end + 1..];
+            let after_bracket = after_inner.strip_prefix("](")?;
+            let link_end = after_bracket.find(')')?;
+            let total = BADGE_PREFIX_LEN + alt_end + 1 + 1 + src_end + 1 + 2 + link_end + 1;
+            return Some(MdImage {
+                src: src.to_string(),
+                alt: alt.to_string(),
+                consumed: total,
+            });
+        }
+
+        let rest = s.strip_prefix("![")?;
+        let close_bracket = rest.find("](")?;
+        let alt = &rest[..close_bracket];
+        let after = &rest[close_bracket + 2..];
+        let close_paren = after.find(')')?;
+        let src = &after[..close_paren];
+        if src.is_empty() {
+            return None;
+        }
+        let total = 2 + close_bracket + 2 + close_paren + 1;
+        Some(MdImage {
+            src: src.to_string(),
+            alt: alt.to_string(),
+            consumed: total,
+        })
+    }
+
+    /// Alternative to `matches!` macro. Avoids uncovered lines issue caused by
+    /// subregions (`^0`) generated by LLVM in the else branch of `matches!`.
+    macro_rules! assert_variant {
+        ($expr:expr, $pat:pat) => {
+            let val = &$expr;
+            assert!(
+                if let $pat = val { true } else { false },
+                "expected {}, got {:?}",
+                stringify!($pat),
+                val
+            );
+        };
+    }
+    // render_diagram: Maps DrawIo result to RenderedSection
+    #[test]
+    fn render_diagram_drawio_returns_ok_section() {
+        let xml = r#"<mxGraphModel><root><mxCell id="0"/><mxCell id="1" parent="0"/></root></mxGraphModel>"#;
+        let section = render_diagram(&DiagramKind::DrawIo, xml, 0);
+        assert_variant!(
+            section,
+            RenderedSection::Image { .. } | RenderedSection::Error { .. }
+        );
+    }
+
+    // dispatch_renderer: DrawIo branch
+    #[test]
+    fn dispatch_renderer_drawio_returns_result() {
+        let block = DiagramBlock {
+            kind: DiagramKind::DrawIo,
+            source: r#"<mxGraphModel><root><mxCell id="0"/></root></mxGraphModel>"#.to_string(),
+        };
+        let result = dispatch_renderer(&block);
+        assert_variant!(result, DiagramResult::Ok(_) | DiagramResult::Err { .. });
+    }
+
+    // dispatch_renderer: Mermaid branch
+    #[test]
+    fn dispatch_renderer_mermaid_when_no_mmdc_returns_command_not_found() {
+        let block = DiagramBlock {
+            kind: DiagramKind::Mermaid,
+            source: "graph TD; A-->B".to_string(),
+        };
+        let result = dispatch_renderer(&block);
+        assert_variant!(
+            result,
+            DiagramResult::CommandNotFound { .. }
+                | DiagramResult::OkPng(_)
+                | DiagramResult::Err { .. }
+        );
+    }
+
+    // dispatch_renderer: PlantUml branch
+    #[test]
+    fn dispatch_renderer_plantuml_when_no_jar_returns_not_installed() {
+        std::env::set_var("PLANTUML_JAR", "/nonexistent/plantuml.jar");
+        let block = DiagramBlock {
+            kind: DiagramKind::PlantUml,
+            source: "@startuml\nA->B\n@enduml".to_string(),
+        };
+        let result = dispatch_renderer(&block);
+        std::env::remove_var("PLANTUML_JAR");
+        assert_variant!(result, DiagramResult::NotInstalled { .. });
+    }
+
+    // try_rasterize: SVG extraction failure case
+    #[test]
+    fn try_rasterize_returns_error_when_no_svg_in_html() {
+        let kind = DiagramKind::DrawIo;
+        let section = try_rasterize(&kind, "source", "<div>no svg here</div>", 0);
+        assert_variant!(section, RenderedSection::Error { .. });
+    }
+
+    // try_rasterize: Success with valid SVG
+    #[test]
+    fn try_rasterize_returns_image_for_valid_svg() {
+        let kind = DiagramKind::DrawIo;
+        let html = r#"<div class="diagram"><svg width="10" height="10"><rect fill="white" width="10" height="10"/></svg></div>"#;
+        let section = try_rasterize(&kind, "source", html, 0);
+        assert_variant!(
+            section,
+            RenderedSection::Image { .. } | RenderedSection::Error { .. }
+        );
+    }
+
+    // decode_png_to_section: Valid PNG
+    #[test]
+    fn decode_png_to_section_returns_image_for_valid_png() {
+        use image::{ImageBuffer, Rgba};
+        let mut buf = Vec::new();
+        let img = ImageBuffer::from_pixel(2, 2, Rgba([100u8, 150, 200, 255]));
+        img.write_to(&mut std::io::Cursor::new(&mut buf), image::ImageFormat::Png)
+            .unwrap();
+        let section = decode_png_to_section(&DiagramKind::DrawIo, "source", buf, 0);
+        assert_variant!(section, RenderedSection::Image { .. });
+    }
+
+    // decode_png_to_section: Invalid data
+    #[test]
+    fn decode_png_to_section_returns_error_for_invalid_data() {
+        let section = decode_png_to_section(&DiagramKind::DrawIo, "source", b"not png".to_vec(), 0);
+        assert_variant!(section, RenderedSection::Error { .. });
+    }
+
+    // map_diagram_result: Exhaustive test for all variants
+    #[test]
+    fn map_diagram_result_ok_delegates_to_try_rasterize() {
+        let section = map_diagram_result(
+            &DiagramKind::DrawIo,
+            "src",
+            DiagramResult::Ok("<svg width=\"10\" height=\"10\"></svg>".to_string()),
+            0,
+        );
+        assert_variant!(
+            section,
+            RenderedSection::Image { .. } | RenderedSection::Error { .. }
+        );
+    }
+
+    #[test]
+    fn map_diagram_result_ok_png_delegates_to_decode() {
+        use image::{ImageBuffer, Rgba};
+        let mut buf = Vec::new();
+        let img = ImageBuffer::from_pixel(2, 2, Rgba([0u8, 0, 0, 255]));
+        img.write_to(&mut std::io::Cursor::new(&mut buf), image::ImageFormat::Png)
+            .unwrap();
+        let section =
+            map_diagram_result(&DiagramKind::Mermaid, "src", DiagramResult::OkPng(buf), 0);
+        assert_variant!(section, RenderedSection::Image { .. });
+    }
+
+    #[test]
+    fn map_diagram_result_err_maps_to_error_section() {
+        let section = map_diagram_result(
+            &DiagramKind::DrawIo,
+            "src",
+            DiagramResult::Err {
+                source: "src".to_string(),
+                error: "render failed".to_string(),
+            },
+            0,
+        );
+        assert_variant!(section, RenderedSection::Error { .. });
+    }
+
+    #[test]
+    fn map_diagram_result_command_not_found_maps_to_section() {
+        let section = map_diagram_result(
+            &DiagramKind::Mermaid,
+            "src",
+            DiagramResult::CommandNotFound {
+                tool_name: "mmdc".to_string(),
+                install_hint: "npm install".to_string(),
+                source: "src".to_string(),
+            },
+            0,
+        );
+        assert_variant!(section, RenderedSection::CommandNotFound { .. });
+    }
+
+    #[test]
+    fn map_diagram_result_not_installed_maps_to_section() {
+        let section = map_diagram_result(
+            &DiagramKind::PlantUml,
+            "src",
+            DiagramResult::NotInstalled {
+                kind: "PlantUML".to_string(),
+                download_url: "https://example.com".to_string(),
+                install_path: std::path::PathBuf::from("/tmp/plantuml.jar"),
+            },
+            0,
+        );
+        assert_variant!(section, RenderedSection::NotInstalled { .. });
+    }
+
+    // render_diagram_mermaid: Integration test (independent of mmdc presence)
+    #[test]
+    fn render_diagram_mermaid_produces_valid_section() {
+        let section = render_diagram(&DiagramKind::Mermaid, "graph TD; A-->B", 0);
+        // CommandNotFound if mmdc is absent, Image if present
+        assert!(!matches!(section, RenderedSection::Pending { .. }));
+    }
+
+    // poll_renders: Receives results from the background thread and updates sections (L200-206)
+    #[test]
+    fn poll_renders_receives_background_result_and_updates_section() {
+        use std::sync::mpsc;
+        let mut pane = PreviewPane::default();
+
+        // Set Pending section
+        pane.sections = vec![RenderedSection::Pending {
+            kind: "DrawIo".to_string(),
+            source: "src".to_string(),
+            source_lines: 0,
+        }];
+
+        // Create an mpsc channel and set it to render_rx
+        let (tx, rx) = mpsc::channel();
+        pane.render_rx = Some(rx);
+
+        // Send a result from the background thread
+        tx.send(RenderMessage::Section {
+            kind: "DrawIo".to_string(),
+            source: "src".to_string(),
+            section: RenderedSection::Markdown("# Result".to_string()),
+        })
+        .unwrap();
+        // Drop tx so the receiver becomes Disconnected
+        drop(tx);
+
+        // egui Context is required to call poll_renders
+        let ctx = egui::Context::default();
+        pane.poll_renders(&ctx);
+
+        // The section has been updated
+        assert_variant!(pane.sections[0], RenderedSection::Markdown(_));
+        // render_rx is None (Pending removed)
+        assert!(pane.render_rx.is_none());
+    }
+
+    // wait_for_renders: Wait until Pending sections are gone (L224-242)
+    #[test]
+    fn wait_for_renders_blocks_until_all_rendered() {
+        use std::sync::mpsc;
+        let mut pane = PreviewPane::default();
+
+        pane.sections = vec![RenderedSection::Pending {
+            kind: "DrawIo".to_string(),
+            source: "src".to_string(),
+            source_lines: 0,
+        }];
+
+        let (tx, rx) = mpsc::channel();
+        pane.render_rx = Some(rx);
+
+        // Send in another thread
+        std::thread::spawn(move || {
+            std::thread::sleep(std::time::Duration::from_millis(10));
+            let _ = tx.send(RenderMessage::Section {
+                kind: "DrawIo".to_string(),
+                source: "src".to_string(),
+                section: RenderedSection::Markdown("# Done".to_string()),
+            });
+        });
+
+        pane.wait_for_renders();
+
+        // No longer Pending after completion
+        assert!(pane.render_rx.is_none());
+        assert_variant!(pane.sections[0], RenderedSection::Markdown(_));
+    }
+
+    // poll_renders: Do nothing without render_rx (L211-213)
+    #[test]
+    fn poll_renders_without_rx_does_nothing() {
+        let mut pane = PreviewPane::default();
+        // render_rx remains None
+        let ctx = egui::Context::default();
+        pane.poll_renders(&ctx);
+        // OK as long as no crash occurs
+        assert!(pane.render_rx.is_none());
+    }
+
+    // full_render: Starts thread and generates Pending section (L140-149)
+    #[test]
+    fn full_render_with_diagram_creates_pending_section_then_renders() {
+        let mut pane = PreviewPane::default();
+        // Content containing a DrawIo diagram -> evaluates to Pending
+        let source = "# Title\n```drawio\n<mxGraphModel><root></root></mxGraphModel>\n```";
+        let cache = std::sync::Arc::new(katana_platform::InMemoryCacheService::default());
+        pane.full_render(
+            source,
+            std::path::Path::new("/tmp/test.md"),
+            cache,
+            false,
+            4,
+        );
+
+        // render_rx is set (because there is a diagram)
+        assert!(pane.render_rx.is_some());
+
+        // Wait and confirm that there are no crashes
+        pane.wait_for_renders();
+        assert!(pane.render_rx.is_none());
+    }
+
+    // full_render: force=true should clear CommonMarkCache
+    // (We cannot directly observe the opaque cache length, but we verify the reset path doesn't crash)
+    #[test]
+    fn full_render_with_force_true_resets_commonmark_cache() {
+        let mut pane = PreviewPane::default();
+        let source = "![image](https://example.com/test.png)";
+        let cache = std::sync::Arc::new(katana_platform::InMemoryCacheService::default());
+
+        // Initial render (force = false)
+        pane.full_render(
+            source,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            false,
+            4,
+        );
+
+        // Force render (force = true) triggers `self.commonmark_cache = CommonMarkCache::default()`
+        pane.full_render(source, std::path::Path::new("/tmp/test.md"), cache, true, 4);
+        assert!(
+            pane.render_rx.is_none(),
+            "Markdown-only render should not have pending background jobs"
+        );
+    }
+
+    // ── parse_md_image / find_next_image unit tests ──
+
+    #[test]
+    fn parse_md_image_simple_image() {
+        let input = "![alt text](path/to/image.png)";
+        let img = parse_md_image(input).unwrap();
+        assert_eq!(img.src, "path/to/image.png");
+        assert_eq!(img.alt, "alt text");
+        assert_eq!(img.consumed, input.len());
+    }
+
+    #[test]
+    fn parse_md_image_simple_with_file_uri() {
+        let input = "![icon](file:///tmp/icon.png)";
+        let img = parse_md_image(input).unwrap();
+        assert_eq!(img.src, "file:///tmp/icon.png");
+        assert_eq!(img.alt, "icon");
+        assert_eq!(img.consumed, input.len());
+    }
+
+    #[test]
+    fn parse_md_image_badge_pattern() {
+        let input = "[![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](LICENSE)";
+        let img = parse_md_image(input).unwrap();
+        assert_eq!(img.src, "https://img.shields.io/badge/License-MIT-blue.svg");
+        assert_eq!(img.alt, "License: MIT");
+        assert_eq!(img.consumed, input.len());
+    }
+
+    #[test]
+    fn parse_md_image_badge_with_url_link() {
+        let input =
+            "[![CI](https://github.com/org/repo/badge.svg)](https://github.com/org/repo/actions)";
+        let img = parse_md_image(input).unwrap();
+        assert_eq!(img.src, "https://github.com/org/repo/badge.svg");
+        assert_eq!(img.alt, "CI");
+        assert_eq!(img.consumed, input.len());
+    }
+
+    #[test]
+    fn parse_md_image_consumed_allows_continuation() {
+        let input = "[![A](https://a.svg)](link1) ![B](https://b.png)";
+        let first = parse_md_image(input).unwrap();
+        assert_eq!(first.alt, "A");
+        let remainder = &input[first.consumed..];
+        let trimmed = remainder.trim_start();
+        let second = parse_md_image(trimmed).unwrap();
+        assert_eq!(second.alt, "B");
+        assert_eq!(second.src, "https://b.png");
+    }
+
+    #[test]
+    fn parse_md_image_empty_src_returns_none() {
+        assert!(parse_md_image("![alt]()").is_none());
+    }
+
+    #[test]
+    fn parse_md_image_plain_text_returns_none() {
+        assert!(parse_md_image("just plain text").is_none());
+    }
+
+    #[test]
+    fn parse_md_image_incomplete_badge_returns_none() {
+        // Missing closing of outer link
+        assert!(parse_md_image("[![alt](src)](").is_none());
+    }
+
+    #[test]
+    fn find_next_image_simple() {
+        assert_eq!(find_next_image("hello ![img](src)"), Some(6));
+    }
+
+    #[test]
+    fn find_next_image_badge() {
+        assert_eq!(find_next_image("[![badge](url)](link)"), Some(0));
+    }
+
+    #[test]
+    fn find_next_image_badge_before_simple() {
+        // Badge starts at 0, simple at 1 — should return 0
+        assert_eq!(
+            find_next_image("[![badge](url)](link) ![img](src)"),
+            Some(0)
+        );
+    }
+
+    #[test]
+    fn find_next_image_no_image() {
+        assert_eq!(find_next_image("no images here"), None);
+    }
+
+    #[test]
+    fn find_next_image_with_preceding_text() {
+        assert_eq!(
+            find_next_image("text before [![badge](url)](link)"),
+            Some(12)
+        );
+    }
+
+    // ── Coverage Gap Fillers (Task 4) ──
+
+    #[test]
+    fn test_coverage_gap_fillers_rendering_logic() {
+        let mut pane = PreviewPane::default();
+        let cache = std::sync::Arc::new(katana_platform::InMemoryCacheService::default());
+
+        // 1. hit is_http = true and get_memory (L202)
+        // 2. hit DiagramResult::Err and ReduceConcurrency (L237)
+        // Note: is_http is true if the source contains "http://"
+        let source = "```drawio\nhttp://invalidxml\n```";
+        pane.full_render(
+            source,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            false,
+            1,
+        );
+        pane.wait_for_renders();
+        assert!(pane.concurrency_reduction_requested);
+
+        // 3. hit force = true and return None (L207)
+        pane.concurrency_reduction_requested = false;
+        pane.full_render(
+            source,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            true,
+            1,
+        );
+        pane.wait_for_renders();
+
+        // 4. hit cached path (L211) and successful recovery/reserialize (L212, L231)
+        // First, put a valid result in cache
+        let diag_src = "http://graph TD; A-->B";
+        let key = get_cache_key(
+            &std::path::PathBuf::from("/tmp/test.md"),
+            &DiagramKind::Mermaid,
+            diag_src,
+        );
+        let valid_res = DiagramResult::Ok("<s></s>".to_string());
+        let valid_json = serde_json::to_string(&valid_res).unwrap();
+        cache.set_memory(&key, valid_json);
+
+        let source2 = format!("```mermaid\n{diag_src}\n```"); // is_http=true
+        pane.full_render(
+            &source2,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            false,
+            1,
+        );
+        pane.wait_for_renders();
+
+        // 5. hit cached path but JSON error -> re-render (L214-222)
+        cache.set_memory(&key, "invalid json".to_string());
+        pane.full_render(
+            &source2,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            false,
+            1,
+        );
+        pane.wait_for_renders();
+
+        // 6. corrupt cache + render error -> hit L223 ReduceConcurrency
+        let diag_src3 = "invalid drawio";
+        let key3 = get_cache_key(
+            &std::path::PathBuf::from("/tmp/test.md"),
+            &katana_core::markdown::diagram::DiagramKind::DrawIo,
+            diag_src3,
+        );
+        let _ = cache.set_persistent(&key3, "invalid json".to_string());
+        let source3 = format!("```drawio\n{diag_src3}\n```");
+        pane.full_render(
+            &source3,
+            std::path::Path::new("/tmp/test.md"),
+            cache.clone(),
+            false,
+            1,
+        );
+        pane.wait_for_renders();
+    }
+
+    #[test]
+    fn cache_key_differs_by_theme() {
+        use katana_core::markdown::color_preset::DiagramColorPreset;
+
+        let path = std::path::Path::new("/tmp/test.md");
+        let kind = DiagramKind::Mermaid;
+        let source = "graph TD; A-->B";
+
+        // Generate key in dark mode
+        DiagramColorPreset::set_dark_mode(true);
+        let key_dark = get_cache_key(path, &kind, source);
+
+        // Generate key in light mode
+        DiagramColorPreset::set_dark_mode(false);
+        let key_light = get_cache_key(path, &kind, source);
+
+        // Restore default
+        DiagramColorPreset::set_dark_mode(true);
+
+        assert_ne!(
+            key_dark, key_light,
+            "Cache key must differ between dark and light themes"
+        );
+    }
+
+    #[test]
+    fn test_coverage_gap_filler_render_message_processing() {
+        let mut pane = PreviewPane::default();
+        let (tx, rx) = std::sync::mpsc::channel();
+        pane.render_rx = Some(rx);
+
+        // hit RenderMessage::ReduceConcurrency in poll_renders (L309-311)
+        tx.send(RenderMessage::ReduceConcurrency).unwrap();
+        let ctx = egui::Context::default();
+        pane.poll_renders(&ctx);
+        assert!(pane.concurrency_reduction_requested);
+
+        // hit RenderMessage::ReduceConcurrency in wait_for_renders (L341-343)
+        pane.concurrency_reduction_requested = false;
+        let (tx, rx) = std::sync::mpsc::channel();
+        pane.render_rx = Some(rx);
+        tx.send(RenderMessage::ReduceConcurrency).unwrap();
+        pane.wait_for_renders();
+        assert!(pane.concurrency_reduction_requested);
+    }
+
+    #[test]
+    fn test_image_preload_queue_processing() {
+        let mut pane = PreviewPane::default();
+        let path = std::path::PathBuf::from("/tmp/test.png");
+        pane.image_preload_queue.push(path.clone());
+
+        let ctx = egui::Context::default();
+        pane.poll_renders(&ctx);
+
+        assert!(pane.image_preload_queue.is_empty());
+        assert!(pane.image_cache.contains(&path));
+    }
+
+    // ── ViewerState tests ──
+
+    #[test]
+    fn viewer_state_default_is_zoom_1_pan_zero() {
+        let state = ViewerState::default();
+        assert_eq!(state.zoom, 1.0);
+        assert_eq!(state.pan, egui::Vec2::ZERO);
+    }
+
+    #[test]
+    fn viewer_state_zoom_in_increases_by_step() {
+        let mut state = ViewerState::default();
+        state.zoom_in();
+        assert_eq!(state.zoom, 1.25);
+    }
+
+    #[test]
+    fn viewer_state_zoom_out_decreases_by_step() {
+        let mut state = ViewerState::default();
+        state.zoom_out();
+        assert_eq!(state.zoom, 0.75);
+    }
+
+    #[test]
+    fn viewer_state_zoom_in_clamps_at_max() {
+        let mut state = ViewerState::default();
+        for _ in 0..20 {
+            state.zoom_in();
+        }
+        assert_eq!(state.zoom, 4.0);
+    }
+
+    #[test]
+    fn viewer_state_zoom_out_clamps_at_min() {
+        let mut state = ViewerState::default();
+        for _ in 0..20 {
+            state.zoom_out();
+        }
+        assert_eq!(state.zoom, 0.25);
+    }
+
+    #[test]
+    fn viewer_state_pan_up() {
+        let mut state = ViewerState::default();
+        state.pan_up();
+        assert_eq!(state.pan, egui::vec2(0.0, -50.0));
+    }
+
+    #[test]
+    fn viewer_state_pan_down() {
+        let mut state = ViewerState::default();
+        state.pan_down();
+        assert_eq!(state.pan, egui::vec2(0.0, 50.0));
+    }
+
+    #[test]
+    fn viewer_state_pan_left() {
+        let mut state = ViewerState::default();
+        state.pan_left();
+        assert_eq!(state.pan, egui::vec2(-50.0, 0.0));
+    }
+
+    #[test]
+    fn viewer_state_pan_right() {
+        let mut state = ViewerState::default();
+        state.pan_right();
+        assert_eq!(state.pan, egui::vec2(50.0, 0.0));
+    }
+
+    #[test]
+    fn viewer_state_pan_by_accumulates() {
+        let mut state = ViewerState::default();
+        state.pan_by(egui::vec2(10.0, 20.0));
+        state.pan_by(egui::vec2(5.0, -10.0));
+        assert_eq!(state.pan, egui::vec2(15.0, 10.0));
+    }
+
+    #[test]
+    fn viewer_state_reset_restores_defaults() {
+        let mut state = ViewerState::default();
+        state.zoom_in();
+        state.zoom_in();
+        state.pan_right();
+        state.pan_down();
+        state.reset();
+        assert_eq!(state, ViewerState::default());
+    }
+
+    #[test]
+    fn preview_pane_viewer_states_default_empty() {
+        let pane = PreviewPane::default();
+        assert!(pane.viewer_states.is_empty());
+        assert!(pane.fullscreen_image.is_none());
+    }
+
+    #[test]
+    fn handle_fullscreen_request_sets_index_for_valid_image() {
+        let mut pane = PreviewPane::default();
+        pane.sections.push(RenderedSection::Image {
+            svg_data: katana_core::markdown::svg_rasterize::RasterizedSvg {
+                width: 1,
+                height: 1,
+                rgba: vec![0; 4],
+            },
+            alt: String::new(),
+            source_lines: 0,
+        });
+        pane.handle_fullscreen_request(Some(0), None);
+        assert_eq!(pane.fullscreen_image, Some(0));
+    }
+
+    #[test]
+    fn handle_fullscreen_request_clears_for_out_of_bounds_index() {
+        let mut pane = PreviewPane::default();
+        pane.handle_fullscreen_request(Some(99), None);
+        assert!(pane.fullscreen_image.is_none());
+    }
+
+    #[test]
+    fn handle_fullscreen_request_clears_for_non_image_section() {
+        let mut pane = PreviewPane::default();
+        pane.sections
+            .push(RenderedSection::Markdown("# Hello".to_string()));
+        pane.handle_fullscreen_request(Some(0), None);
+        assert!(pane.fullscreen_image.is_none());
+    }
+
+    #[test]
+    fn handle_fullscreen_request_noop_on_none() {
+        let mut pane = PreviewPane::default();
+        pane.handle_fullscreen_request(None, None);
+        assert!(pane.fullscreen_image.is_none());
+    }
+
+    #[test]
+    fn handle_fullscreen_request_clears_stale_index() {
+        let mut pane = PreviewPane::default();
+        pane.sections.push(RenderedSection::Image {
+            svg_data: katana_core::markdown::svg_rasterize::RasterizedSvg {
+                width: 1,
+                height: 1,
+                rgba: vec![0; 4],
+            },
+            alt: String::new(),
+            source_lines: 0,
+        });
+        pane.fullscreen_image = Some(0);
+        // Remove section, then validate state.
+        pane.sections.clear();
+        pane.handle_fullscreen_request(None, None);
+        assert!(pane.fullscreen_image.is_none());
+    }
+
+    #[test]
+    fn fullscreen_viewer_state_is_independent() {
+        let mut pane = PreviewPane::default();
+        // Modify main viewer state.
+        pane.viewer_states.push(ViewerState::default());
+        pane.viewer_states[0].zoom_in();
+        // Fullscreen state should remain at defaults.
+        assert!((pane.fullscreen_viewer_state.zoom - 1.0).abs() < f32::EPSILON);
+        assert_eq!(pane.fullscreen_viewer_state.pan, egui::Vec2::ZERO);
+    }
+
+    #[test]
+    fn fullscreen_viewer_state_resets_on_close() {
+        let mut pane = PreviewPane::default();
+        // Simulate zoom in fullscreen.
+        pane.fullscreen_viewer_state.zoom_in();
+        pane.fullscreen_viewer_state.pan_right();
+        assert!(pane.fullscreen_viewer_state.zoom > 1.0);
+        // Close fullscreen should reset state.
+        pane.fullscreen_viewer_state.reset();
+        assert!((pane.fullscreen_viewer_state.zoom - 1.0).abs() < f32::EPSILON);
+        assert_eq!(pane.fullscreen_viewer_state.pan, egui::Vec2::ZERO);
+    }
+
+    #[test]
+    fn i18n_diagram_controller_fields_exist() {
+        use crate::i18n;
+        i18n::set_language("en");
+        let msgs = i18n::get();
+        let dc = &msgs.preview.diagram_controller;
+        assert!(!dc.pan_up.is_empty());
+        assert!(!dc.pan_down.is_empty());
+        assert!(!dc.pan_left.is_empty());
+        assert!(!dc.pan_right.is_empty());
+        assert!(!dc.zoom_in.is_empty());
+        assert!(!dc.zoom_out.is_empty());
+        assert!(!dc.reset.is_empty());
+        assert!(!dc.fullscreen.is_empty());
+        assert!(!dc.close.is_empty());
+    }
+
+    #[test]
+    fn i18n_diagram_controller_ja() {
+        use crate::i18n;
+        i18n::set_language("ja");
+        let msgs = i18n::get();
+        let dc = &msgs.preview.diagram_controller;
+        assert_eq!(dc.pan_up, "\u{4e0a}\u{3078}\u{79fb}\u{52d5}");
+        assert_eq!(dc.reset, "\u{521d}\u{671f}\u{4f4d}\u{7f6e}\u{30fb}\u{30b5}\u{30a4}\u{30ba}\u{306b}\u{30ea}\u{30bb}\u{30c3}\u{30c8}");
+        assert_eq!(dc.fullscreen, "\u{5168}\u{753b}\u{9762}\u{8868}\u{793a}");
+        // Restore default language.
+        i18n::set_language("en");
+    }
+
+    #[test]
+    fn test_fullscreen_viewer_state_apply_result() {
+        let mut pane = PreviewPane::default();
+        let ctx = egui::Context::default();
+        pane.fullscreen_image = Some(0);
+        pane.fullscreen_viewer_state.zoom_in();
+        assert_ne!(pane.fullscreen_viewer_state.zoom, 1.0);
+
+        // Simulate closing modal.
+        pane.apply_fullscreen_result(None, Some(&ctx));
+        assert_eq!(pane.fullscreen_image, None);
+        assert_eq!(pane.fullscreen_viewer_state.zoom, 1.0); // Resets.
+
+        // Simulate keeping modal open.
+        pane.fullscreen_image = Some(0);
+        pane.fullscreen_viewer_state.zoom_in();
+        pane.apply_fullscreen_result(Some(0), Some(&ctx));
+        assert_eq!(pane.fullscreen_image, Some(0));
+        assert_ne!(pane.fullscreen_viewer_state.zoom, 1.0); // Does NOT reset.
+    }
+
+    #[test]
+    fn test_handle_fullscreen_request_context_logic() {
+        let mut pane = PreviewPane::default();
+        let ctx = egui::Context::default();
+
+        pane.sections.push(RenderedSection::Image {
+            svg_data: katana_core::markdown::svg_rasterize::RasterizedSvg {
+                width: 1,
+                height: 1,
+                rgba: vec![0; 4],
+            },
+            alt: "Diagram A".to_string(),
+            source_lines: 0,
+        });
+
+        // Simulating opening modal when it was None
+        pane.handle_fullscreen_request(Some(0), Some(&ctx));
+        assert_eq!(pane.fullscreen_image, Some(0));
+        assert!(!pane.was_os_fullscreen_before_modal); // context default is window mode
+    }
+
+    #[test]
+    fn test_viewer_state_debug() {
+        let state = ViewerState::default();
+        let debug_str = format!("{:?}", state);
+        assert!(debug_str.contains("ViewerState"));
+    }
+    #[test]
+    fn poll_renders_clears_is_loading_on_disconnect() {
+        let mut pane = PreviewPane::default();
+        let (tx, rx) = std::sync::mpsc::channel::<RenderMessage>();
+        drop(tx); // EXPLICIT DISCONNECT
+        pane.render_rx = Some(rx);
+        pane.is_loading = true;
+
+        let ctx = egui::Context::default();
+        pane.poll_renders(&ctx);
+
+        assert!(
+            !pane.is_loading,
+            "poll_renders did not clear is_loading when disconnected"
+        );
+        assert!(
+            pane.render_rx.is_none(),
+            "poll_renders did not drop the disconnected rx"
+        );
+    }
+
+    #[test]
+    fn full_render_sets_is_loading_to_true() {
+        let mut pane = PreviewPane::default();
+        assert!(!pane.is_loading);
+
+        let cache = std::sync::Arc::new(katana_platform::InMemoryCacheService::default());
+        pane.full_render(
+            "```mermaid\ngraph TD;\nA-->B;\n```",
+            std::path::Path::new("test.md"),
+            cache,
+            false,
+            1,
+        );
+
+        assert!(
+            pane.is_loading,
+            "full_render did not set is_loading to true"
+        );
+        assert!(pane.render_rx.is_some());
+    }
+
+    #[test]
+    fn full_render_aborts_on_cancel_token() {
+        let mut pane = PreviewPane::default();
+        let cache = std::sync::Arc::new(katana_platform::InMemoryCacheService::default());
+
+        // Use multiple diagram blocks so that even if the first one races through,
+        // the cancel_token should prevent the rest from completing.
+        let source = concat!(
+            "```mermaid\ngraph TD\nA-->B\n```\n\n",
+            "```mermaid\ngraph TD\nC-->D\n```\n\n",
+            "```mermaid\ngraph TD\nE-->F\n```\n\n",
+            "```mermaid\ngraph TD\nG-->H\n```\n\n",
+            "```mermaid\ngraph TD\nI-->J\n```\n",
+        );
+        pane.full_render(
+            source,
+            &std::path::PathBuf::from("test.md"),
+            cache,
+            true,
+            1, // single-threaded worker to make ordering deterministic
+        );
+        let rx = pane.render_rx.take().unwrap();
+        assert!(pane.is_loading);
+
+        // Immediately cancel before background thread processes all jobs
+        pane.cancel_token
+            .store(true, std::sync::atomic::Ordering::Relaxed);
+
+        // Drain all messages. The first job may have already been picked up
+        // (race), but the cancel_token should prevent the rest.
+        let sections: Vec<_> =
+            std::iter::from_fn(|| rx.recv_timeout(std::time::Duration::from_millis(500)).ok())
+                .filter(|msg| matches!(msg, RenderMessage::Section { .. }))
+                .collect();
+
+        // With 5 diagram jobs and immediate cancel, at most 1 may slip through
+        // due to the race window. The key invariant is that cancel prevents
+        // the majority of work.
+        assert!(
+            sections.len() < 5,
+            "Cancel token should have prevented most renders, but all {} completed",
+            sections.len()
+        );
+    }
+}
